test(qa): extract common methods for cluster scaling tests
