+++ qa/integration-tests/src/test/java/io/camunda/zeebe/it/clustering/dynamic/ScaleUpBrokersTest.java
@@ -7,14 +7,12 @@
  */
 package io.camunda.zeebe.it.clustering.dynamic;
 
-import static io.camunda.zeebe.it.clustering.dynamic.Utils.assertChangeIsPlanned;
+import static io.camunda.zeebe.it.clustering.dynamic.Utils.assertThatAllJobsCanBeCompleted;
+import static io.camunda.zeebe.it.clustering.dynamic.Utils.createInstanceWithAJobOnAllPartitions;
+import static io.camunda.zeebe.it.clustering.dynamic.Utils.scale;
 
 import io.atomix.cluster.MemberId;
 import io.camunda.zeebe.client.ZeebeClient;
-import io.camunda.zeebe.client.api.response.ActivatedJob;
-import io.camunda.zeebe.model.bpmn.Bpmn;
-import io.camunda.zeebe.protocol.Protocol;
-import io.camunda.zeebe.qa.util.actuator.ClusterActuator;
 import io.camunda.zeebe.qa.util.cluster.TestCluster;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.cluster.TestZeebePort;
@@ -26,10 +24,6 @@ import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
-import java.util.stream.IntStream;
-import org.assertj.core.api.Assertions;
-import org.awaitility.Awaitility;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -78,11 +72,12 @@ final class ScaleUpBrokersTest {
     final int newClusterSize = currentClusterSize + 1;
     final int newBrokerId = newClusterSize - 1;
 
-    final var processInstanceKeys = createInstanceWithAJobOnAllPartitions();
+    final var processInstanceKeys =
+        createInstanceWithAJobOnAllPartitions(zeebeClient, JOB_TYPE, PARTITIONS_COUNT);
 
     // when
     createNewBroker(newClusterSize, newBrokerId);
-    scale(newClusterSize);
+    scale(cluster, newClusterSize);
 
     // then
     // verify partition 2 is moved from broker 0 to 1
@@ -95,21 +90,7 @@ final class ScaleUpBrokersTest {
     cluster.awaitCompleteTopology(newClusterSize, 3, 1, Duration.ofSeconds(10));
 
     // then - verify the cluster can still process
-    assertThatAllJobsCanBeCompleted(processInstanceKeys);
-  }
-
-  private void scale(final int newClusterSize) {
-    final var actuator = ClusterActuator.of(cluster.availableGateway());
-    final var newBrokerSet = IntStream.range(0, newClusterSize).boxed().toList();
-
-    // when
-    final var response = actuator.scaleBrokers(newBrokerSet);
-    assertChangeIsPlanned(response);
-
-    // then - verify topology changes are completed
-    Awaitility.await()
-        .timeout(Duration.ofMinutes(2))
-        .untilAsserted(() -> ClusterActuatorAssert.assertThat(cluster).hasAppliedChanges(response));
+    assertThatAllJobsCanBeCompleted(processInstanceKeys, zeebeClient, JOB_TYPE);
   }
 
   @Test
@@ -121,75 +102,19 @@ final class ScaleUpBrokersTest {
 
     // scale to clusterSize 2
     createNewBroker(clusterSize2, broker2);
-    scale(clusterSize2);
+    scale(cluster, clusterSize2);
 
     // when - scale to clusterSize 3
     final int broker3 = broker2 + 1;
     final int finalClusterSize = clusterSize2 + 1;
     createNewBroker(finalClusterSize, broker3);
-    scale(finalClusterSize);
+    scale(cluster, finalClusterSize);
 
     // then -- partition 3 must be moved to new broker
     ClusterActuatorAssert.assertThat(cluster).brokerHasPartition(broker3, 3);
 
     // Changes are reflected in the topology returned by grpc query
     cluster.awaitCompleteTopology(finalClusterSize, 3, 1, Duration.ofSeconds(10));
-  }
-
-  private void assertThatAllJobsCanBeCompleted(final List<Long> processInstanceKeys) {
-    final var jobs =
-        zeebeClient
-            .newActivateJobsCommand()
-            .jobType(JOB_TYPE)
-            .maxJobsToActivate(2 * processInstanceKeys.size())
-            .send()
-            .join();
-
-    Assertions.assertThat(jobs.getJobs().stream().map(ActivatedJob::getProcessInstanceKey).toList())
-        .describedAs("Jobs from all partitions can be activated")
-        .containsExactlyInAnyOrderElementsOf(processInstanceKeys);
-    final var jobCompleteFutures =
-        jobs.getJobs().stream()
-            .map(job -> zeebeClient.newCompleteCommand(job).send().toCompletableFuture())
-            .toList();
-    Assertions.assertThat(
-            CompletableFuture.allOf(jobCompleteFutures.toArray(CompletableFuture[]::new)))
-        .describedAs("All jobs can be completed")
-        .succeedsWithin(Duration.ofSeconds(2));
-  }
-
-  private List<Long> createInstanceWithAJobOnAllPartitions() {
-    final var process =
-        Bpmn.createExecutableProcess("processId")
-            .startEvent()
-            .serviceTask("task", t -> t.zeebeJobType(JOB_TYPE))
-            .endEvent()
-            .done();
-    zeebeClient.newDeployResourceCommand().addProcessModel(process, "process.bpmn").send().join();
-
-    final List<Long> createdProcessInstances = new ArrayList<>();
-    Awaitility.await("Process instances are created in all partitions")
-        .ignoreExceptions() // Might throw exception when a partition has not yet received
-        // deployment distribution
-        .until(
-            () -> {
-              final var result =
-                  zeebeClient
-                      .newCreateInstanceCommand()
-                      .bpmnProcessId("processId")
-                      .latestVersion()
-                      .send()
-                      .join();
-              createdProcessInstances.add(result.getProcessInstanceKey());
-              // repeat until all partitions have atleast one process instance
-              return createdProcessInstances.stream()
-                      .map(Protocol::decodePartitionId)
-                      .distinct()
-                      .count()
-                  == PARTITIONS_COUNT;
-            });
-
-    return createdProcessInstances;
   }
 
   private void createNewBroker(final int newClusterSize, final int newBrokerId) {
